Removed Name property from texture related things. Fixed a bug in AssetManager where loading from packages didn't work.

+++ Hatzap/Assets/AssetPackageHeader.cs
@@ -28,7 +28,7 @@ namespace Hatzap.Assets
                     {
                         Path = br.ReadString(),
                         Position = br.ReadInt64(),
-                        Size = br.ReadInt32()
+                        Size = br.ReadInt64()
                     });
                 }
             }
